\name{CoverageView-internal}
\alias{CoverageView-internal}
\alias{.bed.to.GRanges}
\alias{.gff.to.GRanges}
\alias{.cov.bin}
\alias{cov.bin}
\alias{cov.bin-methods}
\alias{cov.bin,ANY,missing,numeric,GRanges,missing-method}
\alias{cov.bin,ANY,missing,numeric,missing,missing-method}
\alias{cov.bin,ANY,numeric,missing,GRanges,missing-method}
\alias{cov.bin,ANY,numeric,missing,missing,missing-method}
\alias{cov.bin,missing,missing,missing,GRanges,character-method}
\alias{cov.bin,missing,missing,missing,missing,character-method}
\alias{cov.bin,ANY,missing,numeric,missing,CoverageBamFile,missing-method}
\alias{cov.bin,ANY,missing,numeric,missing,CoverageBigWigFile,missing-method}
\alias{cov.bin,ANY,numeric,missing,missing,CoverageBamFile,missing-method}
\alias{cov.bin,ANY,numeric,missing,missing,CoverageBigWigFile,missing-method}
\alias{cov.bin,missing,missing,missing,GRanges,missing,character-method}
\alias{cov.bin,missing,missing,missing,missing,CoverageBamFile,character-method}
\alias{.do.normalization}
\alias{imageScale}

\title{Internal Methods}
\description{
Internal CoverageView functions
 }
\details{
These are not to be called by the user
}
\keyword{ internal }