\name{CoverageView-package}
\alias{CoverageView-package}
\alias{CoverageView}
\docType{package}
\title{
Coverage visualization package for R
}
\description{
Package for the generation of coverage profiles using data originated in different types of NGS experiments
}
\details{
This package provides a framework for the visualization of genome coverage profiles.

It can be used for ChIP-seq experiments, but it can be also used for nucleosome or RNA-seq experiments or whatever type of experiment where one can produce an aligment in the BAM or WIG/BigWIG formats
}
\references{
BAM format specification: http://samtools.sourceforge.net/SAMv1.pdf

WIG format specification: https://genome.ucsc.edu/FAQ/FAQformat.html
}
\author{
Ernesto Lowy

Maintainer: Ernesto Lowy <ernestolowy@gmail.com>
}
